requirements.txt
groq
requests
python-dotenv
py-clob-client